pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["tatsuki"]
packages = [
    { include = "utils"},
    { include = "notebooks"},
    { include = "data"},
    { include = "test"},
    { include = "src"},
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.3"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
hydra-core = "^1.3.2"
jupyterlab = "^3.1.16"
joblib = "^1.2.0"
pytest = "^7.4.0"
optuna = "^3.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 79
max-complexity = 18
ignore = "E203,E266,W503,"

[tool.black]
target-version = ['py39']
line-length = 79
